updates for tests
