+++ csa-runner-app/src/test/java/com/farao_community/farao/swe_csa/app/dichotomy/SweCsaDichotomyRunnerTest.java
@@ -1,0 +1,2 @@
+package com.farao_community.farao.swe_csa.app.dichotomy;public class SweCsaDichotomyRunnerTest {
+}
