+++ csa-runner-app/src/test/java/com/farao_community/farao/swe_csa/app/dichotomy/shifter/LinearScalerTest.java
@@ -1,0 +1,11 @@
+package com.farao_community.farao.swe_csa.app.dichotomy.shifter;
+
+import org.junit.jupiter.api.Test;
+
+public class LinearScalerTest {
+
+    @Test
+    void shiftNetworkTest() {
+
+    }
+}
